Move tests related to containers in typing to test_typing.py
[ghstack-poisoned]
